projects/retail-simulator/pos/requirements.txt
Faker==9.3.1
psycopg2-binary==2.9.1
